[TASK] Move `translate()` function to `ContextService`
