+++ Classes/Service/ContextService.php
@@ -18,6 +18,7 @@ use Romm\Formz\Service\Traits\ExtendedFacadeInstanceTrait;
 use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Service\EnvironmentService;
+use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 
 class ContextService implements SingletonInterface
 {
@@ -70,6 +71,28 @@ class ContextService implements SingletonInterface
         }
 
         return $languageKey;
+    }
+
+    /**
+     * Translation handler. Does the same job as Extbase translation tools,
+     * expect that if the index to the LLL reference is not found, the index is
+     * returned (Extbase would have returned an empty string).
+     *
+     * @param    string $index        The index to the LLL reference.
+     * @param    string $extensionKey Key of the extension containing the LLL reference.
+     * @param    array  $arguments    Arguments passed over to vsprintf.
+     * @return   string               The translated string.
+     */
+    public function translate($index, $extensionKey = null, $arguments = null)
+    {
+        $extensionKey = ($extensionKey) ?: ExtensionService::get()->getExtensionKey();
+        $result = LocalizationUtility::translate($index, $extensionKey, $arguments);
+
+        if ($result === '' && $index !== '') {
+            $result = $index;
+        }
+
+        return $result;
     }
 
     /**
